requirements.txt
python-telegram-bot[job-queue]==20.8
telethon==1.34.0
supabase==2.3.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
beautifulsoup4==4.12.3
cryptg
tqdm
